# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Auth Service API
  version: 1.0.0
  description: OAuth authentication microservice API
  contact:
    name: Auth Service Team

servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Local development server

paths:
  /auth/login/{provider}:
    get:
      summary: Initiate OAuth login
      description: Redirects user to OAuth provider for authentication
      operationId: initiateLogin
      tags:
        - Authentication
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google]
          description: OAuth provider name
      responses:
        '302':
          description: Redirect to OAuth provider
        '400':
          description: Unsupported provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/callback/{provider}:
    get:
      summary: OAuth callback handler
      description: Handles OAuth callback and returns tokens
      operationId: handleCallback
      tags:
        - Authentication
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google]
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      summary: Logout current session
      description: Revokes current session and access token
      operationId: logout
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logout successful

  /auth/logout/all:
    post:
      summary: Logout all sessions
      description: Revokes all user sessions
      operationId: logoutAll
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '204':
          description: All sessions revoked

  /auth/token/refresh:
    post:
      summary: Refresh access token
      description: Exchange refresh token for new access and refresh tokens
      operationId: refreshToken
      tags:
        - Tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/token/validate:
    post:
      summary: Validate token
      description: Introspect token (for debugging/admin, not for high-throughput paths)
      operationId: validateToken
      tags:
        - Tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateTokenRequest'
      responses:
        '200':
          description: Token validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateTokenResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/token/info:
    get:
      summary: Get token info
      description: Get information about current token
      operationId: getTokenInfo
      tags:
        - Tokens
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/me:
    get:
      summary: Get current user
      description: Get current authenticated user profile
      operationId: getMe
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/sessions:
    get:
      summary: Get user sessions
      description: List all active sessions for current user
      operationId: getSessions
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /.well-known/jwks.json:
    get:
      summary: JSON Web Key Set
      description: Public keys for JWT validation
      operationId: getJwks
      tags:
        - Tokens
      responses:
        '200':
          description: JWKS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWKS'
          headers:
            Cache-Control:
              schema:
                type: string
                example: public, max-age=600

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: Opaque refresh token
        expiresIn:
          type: integer
          description: Access token expiration in seconds
        user:
          $ref: '#/components/schemas/User'

    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer

    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    ValidateTokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string

    ValidateTokenResponse:
      type: object
      properties:
        valid:
          type: boolean
        user:
          $ref: '#/components/schemas/User'

    TokenInfo:
      type: object
      properties:
        userId:
          type: string
        expiresAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time

    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
        picture:
          type: string
          nullable: true
        provider:
          type: string

    Session:
      type: object
      properties:
        id:
          type: string
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time

    JWKS:
      type: object
      properties:
        keys:
          type: array
          items:
            $ref: '#/components/schemas/JWK'

    JWK:
      type: object
      properties:
        kty:
          type: string
        kid:
          type: string
        use:
          type: string
        alg:
          type: string
        n:
          type: string
        e:
          type: string

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string